requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
exceptiongroup==1.1.2
greenlet==2.0.2
iniconfig==2.0.0
packaging==23.1
playwright==1.38.0
pyee==9.0.4
pytest==7.4.2
python-slugify==5.0.2
PyYAML==6.0.1
setuptools==68.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
Unidecode==1.2.0
